Test deletebyId

+++ src/main/java/com/rsginer/springboottodolist/task/service/TaskServiceImpl.java
@@ -22,8 +22,6 @@ public class TaskServiceImpl implements TaskService {
         this.taskRepository = taskRepository;
     }
 
-
-
     public Page<Task> getTasks(AppUser user, Pageable pageable) {
         return this.taskRepository.findByAssignedTo(user, pageable);
     }
@@ -40,12 +38,8 @@ public class TaskServiceImpl implements TaskService {
     public Optional<Task> getById(AppUser user, UUID taskId) throws TaskNotCreatedByAndNotAssignedToForbiddenException {
         var task = taskRepository.findById(taskId);
 
-        if (task.isPresent()) {
-            if (ACLTask.isUserAuthorized(user, task.get())) {
-                return task;
-            }
-
-            throw new TaskNotCreatedByAndNotAssignedToForbiddenException(task.get().getId());
+        if (task.isPresent() && ACLTask.isUserAuthorized(user, task.get())) {
+            return task;
         }
 
         return Optional.empty();
@@ -54,20 +48,16 @@ public class TaskServiceImpl implements TaskService {
     public Optional<Task> updateById(AppUser user, UUID taskId, Task task) throws TaskNotCreatedByAndNotAssignedToForbiddenException {
         var foundTask = taskRepository.findById(taskId);
 
-        if (foundTask.isPresent()) {
-            if (ACLTask.isUserAuthorized(user, foundTask.get())) {
-                foundTask.get().setDescription(task.getDescription());
+        if (foundTask.isPresent() && ACLTask.isUserAuthorized(user, foundTask.get())) {
+            foundTask.get().setDescription(task.getDescription());
 
-                if (!task.getAssignedTo().isEmpty()) {
-                    foundTask.get().setAssignedTo(task.getAssignedTo());
-                }
-
-                var updatedTask = taskRepository.save(foundTask.get());
-
-                return Optional.of(updatedTask);
+            if (!task.getAssignedTo().isEmpty()) {
+                foundTask.get().setAssignedTo(task.getAssignedTo());
             }
 
-            throw new TaskNotCreatedByAndNotAssignedToForbiddenException(foundTask.get().getId());
+            var updatedTask = taskRepository.save(foundTask.get());
+
+            return Optional.of(updatedTask);
         }
 
         return Optional.empty();
@@ -76,16 +66,11 @@ public class TaskServiceImpl implements TaskService {
     public Optional<Task> finishById(AppUser user, UUID taskId) throws TaskNotCreatedByAndNotAssignedToForbiddenException {
         var foundTask = taskRepository.findById(taskId);
 
-        if (foundTask.isPresent()) {
+        if (foundTask.isPresent() && ACLTask.isUserAuthorized(user, foundTask.get())) {
+            foundTask.get().setState(TaskState.FINISHED);
+            var finishedTask = taskRepository.save(foundTask.get());
 
-            if (ACLTask.isUserAuthorized(user, foundTask.get())) {
-                foundTask.get().setState(TaskState.FINISHED);
-                var finishedTask = taskRepository.save(foundTask.get());
-
-                return Optional.of(finishedTask);
-            }
-
-            throw new TaskNotCreatedByAndNotAssignedToForbiddenException(foundTask.get().getId());
+            return Optional.of(finishedTask);
         }
 
         return Optional.empty();
@@ -94,16 +79,11 @@ public class TaskServiceImpl implements TaskService {
     public Optional<Boolean> deleteTaskById(AppUser user, UUID taskId) throws TaskNotCreatedByAndNotAssignedToForbiddenException {
         var task = taskRepository.findById(taskId);
 
-        if (task.isPresent()) {
-
-            if (ACLTask.isUserAuthorized(user, task.get())) {
-                taskRepository.deleteById(taskId);
-                return Optional.of(true);
-            }
-
-            throw new TaskNotCreatedByAndNotAssignedToForbiddenException(taskId);
+        if (task.isPresent() && ACLTask.isUserAuthorized(user, task.get())) {
+            taskRepository.deleteById(taskId);
+            return Optional.of(true);
         }
 
-        return Optional.of(false);
+        return Optional.empty();
     }
 }
